requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6

# Database and vector storage
lancedb==0.3.0
numpy==1.24.3
pandas==2.0.3

# Machine Learning and Embeddings
sentence-transformers==2.2.2
transformers==4.35.2
torch>=1.13.0,<2.0.0
torchaudio>=0.13.0,<2.0.0
openai-whisper==20231117
# Use open_clip instead of clip-by-openai for better compatibility
open_clip_torch==2.20.0
Pillow==10.1.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Spotify API and YouTube for audio
spotipy==2.23.0
yt-dlp==2023.11.16

# HTTP and networking
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
click==8.1.7
rich==13.7.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: For advanced audio processing
scipy==1.11.4
scikit-learn==1.3.2
